pyproject.toml
[project]
name = "transcript-archiver"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ctranslate2>=4.4.0",
    "fastparquet>=2024.11.0",
    "ffmpeg-python>=0.2.0",
    "matplotlib>=3.10.7",
    "nltk>=3.9.2",
    "numpy>=2.2.6",
    "opentimelineio>=0.18.1",
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "pyarrow>=21.0.0",
    "pydantic>=2.12.5",
    "scikit-learn>=1.7.2",
    "scipy>=1.15.3",
    "seaborn>=0.13.2",
    "torchaudio>=2.9.1",
    "vadersentiment>=3.3.2",
    "webvtt-py>=0.5.1",
    "whisperx>=3.4.3",
    "yt-dlp>=2025.10.14",
]
